logging_service/requirements.txt
flask
requests
hazelcast-python-client